Move inventory from ini to yml
This commit changes the way we define the main ansible inventory from using a single `.ini` file to a full `inventory` directory with a `hosts.yml` file.

Using a directory for storing the inventory will come handy in the future when using group/host_vars because we will put them in there.