Add `host_vars` folder in `inventory`
This commit introduce the `host_vars` folder inside the inventory to store all the variables related to one particular host.

This way all the variables related to the proxmox API are automatically loaded when using the `node` host, and all the lxc related variables are loaded when provisioning the `lxc` host.