Use ansible vault to store sensitive variables
In this commit we start using ansible vault to securely store all the sensitive informations present in the variables.

We converted the `host_vars` specific files into folders and insite we placed two kind of files: `vars.yml`, and `vault.yml`.
The vars file contains all the variables that can be stored as plain text, the vault file, on the other hand, contains all the sensitive variables that are encrypted using ansible vault and a key-file kept out of version control.