Use `ansible_managed` in templates
This commit uses the `ansible_managed` variable in all the templates to display a custom warning message in the resulting file telling the user not to modify the file manually.

Note: The use of `ansible_manage` will be deprecated in future versions of ansible, but at the moment there is no alternative, and seems to be still accepted (no warining in ansible run). See [this](https://github.com/ansible/ansible/pull/85075) PR on github for more info.