Caddy: Use template for `Caddyfile`
In this commit we switch from a plain `Caddyfile` to using a jinja template file with variables derermined from the caddy inventory.
This configuration remains simple, but allows us to have greater control over certain values that can change depending on the context; also it is easier to store secrets this way.