Move playbooks inside `playbooks/lxcs`
This commit moves all the playbooks related to specific LXCs inside a the directory `playbooks/lxcs`.

This is done to better separate all the playbooks by the type of machine they are deployed into.
In the future we will add support for docker containers adding a `containers` sub-directory; also we will manage multiple proxmox instances adding playbooks for theirs setups too.