Role/setup-base-lxc: Add shared_data_clients group
In this commit we change the default value for some variables inside the role `setup-base-lxc`.
Those variables are responsible of creating a user and group inside the lxc used to share files across the host and other containers.

Previously the default was to always create such user, but now we switched to not doing that, and all the LXCs that need the user should set `lxc_has_shared_user` to `true` manually.
This enables better code quality.

In this commit we also create a new group called `shared_data_clients` with all the LXC hosts that need a to shared data amongs them. Inside `group_vars/shared_data_clients.yml` we enable the shared group as stated before using some sane values.

Doing all of that will be particuary usefull in the future when we will need to access variables like `lxc_shared_uid/gid` in jinja2 templates.