Add `portfolio` container
This commit adds a playbook to manage the deployment of my personal portfolio website as a docker stack.
This container is different from the other since there is not an public image available in a registry, but we need to clone the whole git project and build the image localy.