shared_data_clients: Fix missing node in group
This commit adds the `node` Ansible host to the `shared_data_clients` group.
This is essential because the proxmox host must access this variable when creating the LXCs; also in the future we will automate the creation of the shared user in the host and those variables will be accessed anyways.