Add `immich` LXC
This commit adds code to provision an LXC running the immich service.
For better ease of installation the service is run inside docker, so this LXC is basically a glorified docker LXC only for the service.

We choose to install immich in a dedicate docker instance because it is more resource intensive than the other containers (using deep-learning, etc) so we can manage it better.