role/create-lxc: use `community.proxmox`
This commit replace the old, and almost deprecated, `community.general.proxmox` collection in favor of the modern `community.proxmox`.

In this commit we also moved around parts of the tasks.
The `create.yml` task now ensures that the LXC is created correctly and starts it.
The `delete.yml` task stops the LXC only if it is present and running before deleting it.
After this refactor the two task files `start.yml`, and `status.yml` were deleted as not in use anymore.
This refactor fixes a bug present before where the execution will fail when checking the status of a non-existing LXC.