role/create-lxc: Add flag for deleting LXC
Before this commit, every time we wanted to use this role to delete an existing LXC we needed to call this role using `tasks_from:delete.yml` in order to call the `delete.yml` task.

This commit remedies to this problem by adding the `lxc_delete` variable that is used in the `main.yml` task to decide whether to create the LXC or delete it.
This way, deleting an LXC is only the matter of setting this variable to true and the role will do everything on its own.