role/create-lxc: Ensure host bind mounts exists
When creating an LXC we can specify additional mounting points using the variable `lxc_mounts`. These will bind mount to a particular directory on the Proxmox host.
If that folder does not exist the process of creating the LXC will fail with an error.
To avoid this, this commit introduces a task that ensures that all bind mount folders on the host are present.