role/create-lxc: Fix bug obtaining LXC status.
In this commit we fix a bug introduced by the recent feature that allows to make API calls both with password and token authentication.
This feature involves using two tasks, both with a when condition and some common variables, to decide which auth method to use for all the Proxmox API calls.

The bug was present when retrieving the status of a running LXC. The code would register the same variable for both tasks. When running Ansible will register any variable even if the task was skipped.
This means that in the case when we used password authentication (the password task is run before) the variable registered will be overridden by the skipped token task obtaining an invalid status for the LXC.

To fix this bug we split all the tasks needed to retrieve the status into a different file called `get_lxc_status.yml`. These tasks define a `lxc_staus` variable that can be used consistently in both creat and delete flows.