role/create-lxc: Ensure host bind mounts exists
When creating an LXC we can specify additional mounting points using the variable `lxc_mounts`. These will bind mount to a particular directory on the Proxmox host.
If that folder does not exist the process of creating the LXC will fail with an error.

This commit adds a task to ensure that all defined bind mount folders are present on the Proxmox host.