Clean up setup-base-lxc docs
This commit changes the variable definition inside the `README.md` file of the `setup-base-lxc` role.
This commit is part of as series of commits where we change how the roles variables are documents.
Now we use a simple table instead of yaml code blocks.
This ensures that all the available variables are documented clearly and consistently.