Rename template files to include target extention
This commit renames all the template files to include the target extension in the name.
This is done to avoid confusion and mistakes when working with Jinja templates that will be transformed into a variety of different file formats.
Putting the target extension in the name will clear all doubt when opening a template file.

After this change, a file named `file.ext` will have a corresponding template file named `file.ext.j2`.
If the file does not have any extension, the template will only end with `.j2`.
For example, `env.j2` will become `.env`.