Check if docker and compose are present
This commit adds a task to the `docker-compose-deploy` role that ensures that Docker Engine and Docker Compose are installed to the host node before running the role.
In case the required packages are missing the role fails with a custom error message.

Docher and Compose are not installed by this role which only responsability is to deploy a Docker Compose stack.