Use Ansible `path_join` to define paths
This commit define a new `compose_base_dir` variable representing the root path for any compose porject.
The `compose_target_dir` variable was moved to `/vars/main.yml` since it is an internal variable that does not need to be modified by the user of the role.

In this commit we also used `ansible.builtin.path_join` when componing path string insetead of simple templating. This is done to avoid problems with malformed strings.