Add `caddy` Ansible Role
This commit introduce the `caddy` Ansible Role to set up and manage the Caddy reverse proxy.
Installation of the service is delegated to the `maxhoesel.caddy` Ansible Collection, while further configuration is performed by this role.

This Role includes features like: per-site configuration stored in separate files with automatical cleanup of old sites, and support for both ACME and local certificates.

This commit also replace the old manual caddy configuration in the playbooks with this new role.