Fix validate command in `system-users` role
The `system-users` Ansible Role edits the sudoers file in oreder to include options for users and groups. For this reason, the file must be validated using `visudo` as per Linux convention.

In one instance the validate the path to `visudo` contained a typo resulting in a failed execution. This commit replaces the hard-coded paths with an Ansible variable to avoid future errors.