Use `__` (dunder) for role-private variables
All variables with a private scope in an Ansible Role should use the `__` (dunder) prefix to indicate that their value is not important. This is similar to how python treats private class members, and follows Ansible guidelines.

This commit renames all variables inside roles created using the `register` keyword or the `ansible.builtin.set_fact` module to use this prefix.