Add `role_name_` suffix to role-variables
As per the `var-naming[no-role-prefix]` Ansible lint rule, variables names from within roles should use `role_name_` as a prefix. Underlines are accepted before the prefix.
This commit adds this suffix to some variables that were not using it. In some roles it is not possible to use the suffix in all variables, for this reason they are marked inside `.ansible-lint-ignore`.